test(server): implementing jest setupFiles
