+++ server/jest.config.js
@@ -9,6 +9,7 @@ exports.default = {
 	clearMocks: true,
 	// Indicates whether the coverage information should be collected while executing the test
 	collectCoverage: true,
+	// An array of glob patterns indicating a set of files for which coverage information should be collected
 	collectCoverageFrom: [
 		'src/**/*.ts',
 		'!src/migrations/**',
@@ -28,6 +29,8 @@ exports.default = {
 	coveragePathIgnorePatterns: ['/node_modules/', '/src/migrations/'],
 	// Indicates which provider should be used to instrument code for coverage
 	coverageProvider: 'v8',
+	// A list of files that Jest should use to before all test suites
+	setupFilesAfterEnv: ['<rootDir>/tests/setup.jest.ts'],
 	// A preset that is used as a base for Jest's configuration
 	preset: 'ts-jest',
 	// The test environment that will be used for testing
