+++ server/tests/integrations/Authentication.integration.test.ts
@@ -1,29 +1,12 @@
 import { MemberServices } from 'services/MemberServices';
 import { SessionServices } from 'services/SessionServices';
 
-import { Database } from 'db';
 import { ErrorMessages } from 'utils/enums/ErrorMessages';
 import { randomBytes } from 'crypto';
 
 describe('Authentication integration test', () => {
 	const MemberService = new MemberServices();
 	const SessionService = new SessionServices();
-
-	beforeAll(async () => {
-		jest.spyOn(console, 'info').mockImplementation(() => {});
-		await Database.start();
-	});
-
-	afterAll(async () => {
-		await Database.stop();
-	});
-
-	beforeEach(async () => {
-		for (const entity of Database.entityMetadatas()) {
-			const repository = Database.repository(entity.name);
-			await repository.query(`TRUNCATE ${entity.tableName} RESTART IDENTITY CASCADE;`);
-		}
-	});
 
 	describe("when email address doesn't belong to existing user", () => {
 		it('throws invalid credentials error', async () => {
