+++ server/tests/integrations/services/BaseServices/create.integration.test.ts
@@ -5,22 +5,6 @@ import { MemberServices } from 'services/MemberServices';
 describe('Create integration test', () => {
 	const MemberService = new MemberServices();
 
-	beforeAll(async () => {
-		jest.spyOn(console, 'info').mockImplementation(() => {});
-		await Database.start();
-	});
-
-	afterAll(async () => {
-		await Database.stop();
-	});
-
-	beforeEach(async () => {
-		for (const entity of Database.entityMetadatas()) {
-			const repository = Database.repository(entity.name);
-			await repository.query(`TRUNCATE ${entity.tableName} RESTART IDENTITY CASCADE;`);
-		}
-	});
-
 	describe('when try create an user with valid data', () => {
 		describe('when trying to create an element that already exists', () => {
 			beforeEach(() => {
