+++ server/tests/integrations/services/BaseServices/delete.integration.test.ts
@@ -6,22 +6,6 @@ import { v4 as uuid } from 'uuid';
 describe('Delete integration test', () => {
 	const MemberService = new MemberServices();
 
-	beforeAll(async () => {
-		jest.spyOn(console, 'info').mockImplementation(() => {});
-		await Database.start();
-	});
-
-	afterAll(async () => {
-		await Database.stop();
-	});
-
-	beforeEach(async () => {
-		for (const entity of Database.entityMetadatas()) {
-			const repository = Database.repository(entity.name);
-			await repository.query(`TRUNCATE ${entity.tableName} RESTART IDENTITY CASCADE;`);
-		}
-	});
-
 	describe('when delete a element', () => {
 		describe('when id is invalid', () => {
 			it('throw an error not found', async () => {
