+++ server/tests/integrations/services/BaseServices/findBy.integration.test.ts
@@ -4,22 +4,6 @@ import { MemberServices } from 'services/MemberServices';
 describe('FindBy integration test', () => {
 	const MemberService = new MemberServices();
 
-	beforeAll(async () => {
-		jest.spyOn(console, 'info').mockImplementation(() => {});
-		await Database.start();
-	});
-
-	afterAll(async () => {
-		await Database.stop();
-	});
-
-	beforeEach(async () => {
-		for (const entity of Database.entityMetadatas()) {
-			const repository = Database.repository(entity.name);
-			await repository.query(`TRUNCATE ${entity.tableName} RESTART IDENTITY CASCADE;`);
-		}
-	});
-
 	describe('when no records exist', () => {
 		it('returns an empty array', async () => {
 			const email = 'unknow@test.com';
