+++ server/tests/integrations/services/BaseServices/findById.integration.test.ts
@@ -5,22 +5,6 @@ import { v4 as uuid } from 'uuid';
 describe('FindById integration test', () => {
 	const MemberService = new MemberServices();
 
-	beforeAll(async () => {
-		jest.spyOn(console, 'info').mockImplementation(() => {});
-		await Database.start();
-	});
-
-	afterAll(async () => {
-		await Database.stop();
-	});
-
-	beforeEach(async () => {
-		for (const entity of Database.entityMetadatas()) {
-			const repository = Database.repository(entity.name);
-			await repository.query(`TRUNCATE ${entity.tableName} RESTART IDENTITY CASCADE;`);
-		}
-	});
-
 	describe('when no records exist', () => {
 		it('returns an nullable element', async () => {
 			expect(await MemberService.findById(uuid())).toBeNull();
