+++ server/tests/integrations/services/MemberServices/deleteAccount.integration.test.ts
@@ -6,22 +6,6 @@ import { ErrorMessages } from 'utils/enums/ErrorMessages';
 describe('Delete a Member account integration test', () => {
 	const MemberService = new MemberServices();
 
-	beforeAll(async () => {
-		jest.spyOn(console, 'info').mockImplementation(() => {});
-		await Database.start();
-	});
-
-	afterAll(async () => {
-		await Database.stop();
-	});
-
-	beforeEach(async () => {
-		for (const entity of Database.entityMetadatas()) {
-			const repository = Database.repository(entity.name);
-			await repository.query(`TRUNCATE ${entity.tableName} RESTART IDENTITY CASCADE;`);
-		}
-	});
-
 	describe('when delete account with invalid password', () => {
 		it('throw an invalid password error', async () => {
 			const username = 'username';
