+++ server/tests/integrations/services/MemberServices/signUp.integration.test.ts
@@ -6,22 +6,6 @@ import { ErrorMessages } from 'utils/enums/ErrorMessages';
 describe('Singup a Member integration test', () => {
 	const MemberService = new MemberServices();
 
-	beforeAll(async () => {
-		jest.spyOn(console, 'info').mockImplementation(() => {});
-		await Database.start();
-	});
-
-	afterAll(async () => {
-		await Database.stop();
-	});
-
-	beforeEach(async () => {
-		for (const entity of Database.entityMetadatas()) {
-			const repository = Database.repository(entity.name);
-			await repository.query(`TRUNCATE ${entity.tableName} RESTART IDENTITY CASCADE;`);
-		}
-	});
-
 	describe("when email address doesn't belong to existing user", () => {
 		it('return an member', async () => {
 			const data = {
