+++ server/tests/integrations/services/MemberServices/updatePassword.integration.test.ts
@@ -7,22 +7,6 @@ import { ErrorMessages } from 'utils/enums/ErrorMessages';
 describe('Update a Member password integration test', () => {
 	const MemberService = new MemberServices();
 
-	beforeAll(async () => {
-		jest.spyOn(console, 'info').mockImplementation(() => {});
-		await Database.start();
-	});
-
-	afterAll(async () => {
-		await Database.stop();
-	});
-
-	beforeEach(async () => {
-		for (const entity of Database.entityMetadatas()) {
-			const repository = Database.repository(entity.name);
-			await repository.query(`TRUNCATE ${entity.tableName} RESTART IDENTITY CASCADE;`);
-		}
-	});
-
 	describe('when update password with invalid email', () => {
 		it('throw an invalid email error', async () => {
 			const username = 'username';
