+++ server/tests/integrations/services/MemberServices/validEmail.integration.test.ts
@@ -5,22 +5,6 @@ import { Database } from 'db';
 describe('Validate a Member email integration test', () => {
 	const MemberService = new MemberServices();
 
-	beforeAll(async () => {
-		jest.spyOn(console, 'info').mockImplementation(() => {});
-		await Database.start();
-	});
-
-	afterAll(async () => {
-		await Database.stop();
-	});
-
-	beforeEach(async () => {
-		for (const entity of Database.entityMetadatas()) {
-			const repository = Database.repository(entity.name);
-			await repository.query(`TRUNCATE ${entity.tableName} RESTART IDENTITY CASCADE;`);
-		}
-	});
-
 	describe('when valid email', () => {
 		it('return an member', async () => {
 			const username = 'username';
