+++ server/tests/integrations/services/RoutingTokensServices/findByToken.integration.test.ts
@@ -7,22 +7,6 @@ describe('Find a RoutingToken ByToken integration test', () => {
 	const MemberService = new MemberServices();
 	const RoutingTokenService = new RoutingTokenServices();
 
-	beforeAll(async () => {
-		jest.spyOn(console, 'info').mockImplementation(() => {});
-		await Database.start();
-	});
-
-	afterAll(async () => {
-		await Database.stop();
-	});
-
-	beforeEach(async () => {
-		for (const entity of Database.entityMetadatas()) {
-			const repository = Database.repository(entity.name);
-			await repository.query(`TRUNCATE ${entity.tableName} RESTART IDENTITY CASCADE;`);
-		}
-	});
-
 	describe('when find a routing token by the token', () => {
 		it('return an member', async () => {
 			const username = 'username';
